fix imports and database

+++ src/main/java/com/alamin/ecommerce/cart/CartService.java
@@ -53,7 +53,7 @@ public class CartService {
 
     private void updateCartTotals(Cart cart) {
         //int totalQuantity = cart.getItems().stream().mapToInt(CartItem::getQuantity).sum();
-        double totalPrice = cart.getItems().stream().mapToDouble(item -> item.getQuantity() * item.getPrice()).sum();
+        int totalPrice = cart.getItems().stream().mapToInt(item -> item.getQuantity() * item.getPrice()).sum();
         //cart.setTotalQuantity(totalQuantity);
         cart.setTotalPrice(totalPrice);
     }
